% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygon.R
\name{check_for_internal_polygon}
\alias{check_for_internal_polygon}
\title{Check for (and return if so) the shape if it is "internal" (contained within
\code{hpa.spatial}/\code{hpa.spatial.data})}
\usage{
check_for_internal_polygon(
  name = NULL,
  area = NULL,
  year = NULL,
  export_dir = tempdir(),
  quiet = getOption("hpa.spatial.quiet", FALSE),
  ...
)
}
\arguments{
\item{name}{A character string names to identify data not kept on
\code{absmapsdata}.}

\item{area}{A character string names to identify data not kept on
\code{absmapsdata}.}

\item{year}{A character string names to identify data not kept on
\code{absmapsdata}.}

\item{export_dir}{The directory to store the downloaded data.}

\item{quiet}{whether to be quiet about warnings and messages. Set package
level quiet-ness with \code{options(hpa.spatial.quiet = TRUE)}.}

\item{...}{Additional, ignored arguments.}
}
\value{
An \code{sf} object or, if no pkg data found, \code{NULL}.
}
\description{
Check for (and return if so) the shape if it is "internal" (contained within
\code{hpa.spatial}/\code{hpa.spatial.data})
}
\examples{
# Get the Local Hospital Network (LHN) shapes.
shp <- check_for_internal_polygon(name = "LHN")
}
