% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_child_geo.R
\name{create_child_geo}
\alias{create_child_geo}
\title{Create a child geography (polygon) which fits neatly into a parent geography}
\usage{
create_child_geo(
  child_geo,
  parent_geo,
  mb_geo = get_mb21_pop(),
  mb_poly = get_mb21_poly(),
  minimum_majority_portions = 0.95
)
}
\arguments{
\item{child_geo}{Child geography (\code{sf} polygon or multipolygon) which
is to be adjusted to fit within \code{parent_geo}.}

\item{parent_geo}{Parent geography (\code{sf} polygon or multipolygon) which
the child geography is being adapted to fit within.}

\item{mb_geo}{Mesh blocks geography with \code{Person} column for population
based mapping. Defaults to use 2021 edition and census data.}

\item{mb_poly}{Mesh blocks polygon for dissolving within subsets of polygons
from \code{child_geo} that cross multiple polygons in \code{parent_geo}.
Defaults to use 2021 edition.}

\item{minimum_majority_portions}{If the child geography has more than
\code{minimum_majority_portions} proportion of their population within a
single polygon within the \code{parent_geo}, it will not be split into
multiple separate polygons.}
}
\value{
An \code{sf} polygon or multipolygon.
}
\description{
Create a child geography (polygon) which fits neatly into a parent geography
}
\examples{
\donttest{
sa3 <- get_polygon("sa32016")
lhns <- get_polygon("LHN")
create_child_geo(sa3, lhns)
}
}
