% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shapefiles.R
\name{download_shapefiles}
\alias{download_shapefiles}
\title{Download ASGS shape files via \code{ASGS.foyer}.}
\usage{
download_shapefiles(...)
}
\arguments{
\item{...}{Args passed to \code{ASGS.foyer::install_ASGS()}}
}
\value{
\code{NULL}
}
\description{
Download ASGS shape files via \code{ASGS.foyer}.
}
\examples{
\donttest{
# download_shapefiles()
}
}
