% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hpa_spatial_data.R
\name{get_mb21_poly}
\alias{get_mb21_poly}
\title{Get Mesh Blocks (2021 edition) polygons.}
\usage{
get_mb21_poly(export_dir = tempdir())
}
\arguments{
\item{export_dir}{The directory to store the downloaded data.}
}
\value{
a \code{sf} object.
}
\description{
Get Mesh Blocks (2021 edition) polygons.
}
\examples{
get_mb21_poly()
}
