% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygon.R
\name{get_mb21_pop}
\alias{get_mb21_pop}
\title{Get Mesh Blocks (2021 edition) and population counts}
\usage{
get_mb21_pop(export_dir = tempdir())
}
\arguments{
\item{export_dir}{The directory to store the downloaded data.}
}
\value{
An \code{sf} object.
}
\description{
Get Mesh Blocks (2021 edition) and population counts
}
\examples{
get_mb21_pop()
}
