% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygon.R
\name{get_mmm19_poly}
\alias{get_mmm19_poly}
\title{Modified Monash Model polygons (SA1, 2016 edition, polygons with MMM, 2019
edition column)}
\usage{
get_mmm19_poly(export_dir = tempdir())
}
\arguments{
\item{export_dir}{The directory to store the downloaded data.}
}
\value{
An \code{sf} object.
}
\description{
Modified Monash Model polygons (SA1, 2016 edition, polygons with MMM, 2019
edition column)
}
\examples{
get_mmm19_poly()
}
