% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa.spatial-package.R
\docType{package}
\name{hpa.spatial-package}
\alias{hpa.spatial}
\alias{hpa.spatial-package}
\title{hpa.spatial: Data and helpful functions for spatial data analyses for Australian geometries}
\description{
Data and helpful functions for spatial data analyses.
}
\author{
\strong{Maintainer}: Rex Parsons \email{rex.parsons@healthpolicy.com}

}
\keyword{internal}
