% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhn}
\alias{lhn}
\title{\code{sf} Multipolygon data for the Local Hospital Network boundaries for Australia.}
\format{
## `lhn`

A \code{sf} object.


shows the boundaries of the Local Hospital Network in Australia.
}
\source{
<https://hub.arcgis.com/datasets/ACSQHC::local-hospital-networks/explore>
}
\usage{
lhn
}
\description{
\code{sf} Multipolygon data for the Local Hospital Network boundaries for Australia.
}
\keyword{datasets}
