% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data_with_correspondence.R
\name{map_data_with_correspondence}
\alias{map_data_with_correspondence}
\title{Map data between editions using correspondence tables from the ABS.}
\usage{
map_data_with_correspondence(
  codes,
  values,
  from_area,
  from_year,
  to_area,
  to_year,
  value_type = c("units", "aggs"),
  round = FALSE
)
}
\arguments{
\item{codes}{codes representing locations relevant to the \code{from_area}.
SA1 or SA2, for example.}

\item{values}{values associated with codes to be alloocated to newly mapped
codes.}

\item{from_area}{The area you want to correspond FROM (ie the areas your data
are currently in). For example: "sa1", "sa2, "sa3", "sa4".}

\item{from_year}{The year you want to correspond FROM. For example: 2011,
2016.}

\item{to_area}{The area you want to correspond TO (ie the areas you want your
data to be in).}

\item{to_year}{The year you want to correspond TO.}

\item{value_type}{Whether the data are unit level or aggregate level data.
Unit level data is randomly allocated to new locations based on proportions
in the correspondence table, aggregate data is dispersed based on the
proportion across the (potentially multiple) mapped codes.}

\item{round}{Whether or not to round the resulting mapped values to be whole
numbers (maybe be useful when mapping count, aggregate values which may
otherwise return decimal values in the mapped areas).}
}
\value{
A \code{data.frame}.
}
\description{
Map data between editions using correspondence tables from the ABS.
}
\examples{
map_data_with_correspondence(
  codes = c(107011130, 107041149),
  values = c(10, 10),
  from_area = "sa2",
  from_year = 2011,
  to_area = "sa2",
  to_year = 2016,
  value_type = "units"
)
map_data_with_correspondence(
  codes = c(107011130, 107041149),
  values = c(10, 10),
  from_area = "sa2",
  from_year = 2011,
  to_area = "sa2",
  to_year = 2016,
  value_type = "aggs"
)

}
