% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb21_pop}
\alias{mb21_pop}
\title{\code{sf} Coordinates data (centroids) for the meshblocks (2021 edition) for Australia including population counts (\code{Person}).}
\format{
## `mb21_pop`

A \code{sf} object.
}
\source{
Mesh Blocks 2021 shapefile <https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files> and population data <https://www.abs.gov.au/census/guide-census-data/mesh-block-counts/latest-release>.
}
\usage{
mb21_pop
}
\description{
\code{sf} Coordinates data (centroids) for the meshblocks (2021 edition) for Australia including population counts (\code{Person}).
}
\keyword{datasets}
