% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw_lhd}
\alias{nsw_lhd}
\title{\code{sf} Multipolygon data for the local health district boundaries within Queensland..}
\format{
## `nsw_lhd`

A \code{sf} object.


shows the boundaries of the local health districts in NSW.
}
\source{
<<https://www.google.com/maps/d/u/0/viewer?mid=1Dv1JRTGmzlm83tBv7tb8vQcOQXY>
}
\usage{
nsw_lhd
}
\description{
\code{sf} Multipolygon data for the local health district boundaries within Queensland..
}
\keyword{datasets}
