% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phn}
\alias{phn}
\title{\code{sf} Multipolygon data for the Primary Health Network boundaries for Australia.}
\format{
## `phn`

A \code{sf} object.


shows the boundaries of the Primary Health Network in Australia.
}
\source{
<https://data.gov.au/dataset/ds-dga-ef2d28a4-1ed5-47d0-8e3a-46e25bc4f66b/details>
}
\usage{
phn
}
\description{
\code{sf} Multipolygon data for the Primary Health Network boundaries for Australia.
}
\keyword{datasets}
