% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qld_hhs}
\alias{qld_hhs}
\title{\code{sf} Multipolygon data for the hospital and health service boundaries within Queensland..}
\format{
## `qld_hhs`

A \code{sf} object.


shows the boundaries of the hospital and health services in Queensland.
}
\source{
<https://qldspatial.information.qld.gov.au/catalogue/custom/detail.page?fid={A4661F6D-0013-46EE-A446-A45F01A64D46}>
}
\usage{
qld_hhs
}
\description{
\code{sf} Multipolygon data for the hospital and health service boundaries within Queensland..
}
\keyword{datasets}
