% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qld_sa1_2016_to_hhs}
\alias{qld_sa1_2016_to_hhs}
\title{\code{data.frame} A table with mappings from SA1's to HHS in QLD.}
\format{
## `qld_sa1_2016_to_hhs`

A \code{data.frame} object.


A lookup table between SA1 codes and HHS (majority contained).
}
\source{
<hqld_hhs and ASGS-SA1 boundaries>
}
\usage{
qld_sa1_2016_to_hhs
}
\description{
\code{data.frame} A table with mappings from SA1's to HHS in QLD.
}
\keyword{datasets}
