% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_erp23}
\alias{sa2_erp23}
\title{Estimated Residential Population 2023}
\format{
`sa2_erp23`
A data frame with 2,472 rows and 2 columns:
\describe{
  \item{sa2_code}{SA2 code (2021 edition)}
  \item{erp}{estimated number of residents}
}
}
\source{
<https://www.abs.gov.au/statistics/people/population/regional-population/latest-release#data-downloads>
}
\usage{
sa2_erp23
}
\description{
(Estimated) population data at the SA2 level (2021 edition) from the ABS.
}
\keyword{datasets}
