% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa_lhn}
\alias{sa_lhn}
\title{\code{sf} Multipolygon data for the local hospital network boundaries within South Australia.}
\format{
## `sa_lhn`

A \code{sf} object.


shows the boundaries of the local hospital network in SA.
}
\source{
<https://data.gov.au/dataset/ds-sa-120bdc9e-1c96-4ea5-b98c-aa148bb33a10/details?q=primary%20health%20network>
}
\usage{
sa_lhn
}
\description{
\code{sf} Multipolygon data for the local hospital network boundaries within South Australia.
}
\keyword{datasets}
