% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa_phn_correspondence_tables}
\alias{sa_phn_correspondence_tables}
\title{Statistical Areas (ASGS) to PHN area correspondence data.}
\format{
## `sa_phn_correspondence_tables`

A \code{list} of \code{data.frame}s.


contains correspondence data between ASGS structures (SA1, SA2, and SA3) to PHNs across Australia.
}
\source{
<https://www.health.gov.au/resources/collections/primary-health-networks-phns-collection-of-concordance-files>
}
\usage{
sa_phn_correspondence_tables
}
\description{
Statistical Areas (ASGS) to PHN area correspondence data.
}
\keyword{datasets}
