% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygon.R
\name{check_for_internal_polygon}
\alias{check_for_internal_polygon}
\title{check_for_internal_polygon}
\usage{
check_for_internal_polygon(name = NULL, area = NULL, year = NULL, ...)
}
\arguments{
\item{name}{a character string names to identify data not kept on absmapsdata.}

\item{area}{a character string names to identify data not kept on absmapsdata.}

\item{year}{a character string names to identify data not kept on absmapsdata.}

\item{...}{additional, ignored arguments.}
}
\value{
a \code{sf} object or, if no pkg data found, \code{NULL}.
}
\description{
check_for_internal_polygon
}
\examples{
# hospital and health services (HHS) is the name for the local hospital networks in QLD.
shp <- check_for_internal_polygon(name = "HHS")
}
