% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa.spatial-package.R
\docType{package}
\name{hpa.spatial-package}
\alias{hpa.spatial}
\alias{hpa.spatial-package}
\title{hpa.spatial: (Australian) Data Access and Helpful Functions for Spatial Data Analyses}
\description{
Data and helpful functions for spatial data analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://healthpolicyanalysis.github.io/hpa.spatial/}
}

}
\author{
\strong{Maintainer}: Rex Parsons \email{rex.parsons@nousgroup.com}

}
\keyword{internal}
