% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_correspondence_tbl.R
\name{make_correspondence_tbl}
\alias{make_correspondence_tbl}
\title{Create a correspondence table between any two geographies.}
\usage{
make_correspondence_tbl(from_geo, to_geo, mb_geo)
}
\arguments{
\item{from_geo}{an \code{{sf}} POLYGON object.}

\item{to_geo}{an \code{{sf}} POLYGON object.}

\item{mb_geo}{an \code{{sf}} POINT object where the points are the centroids
of a small area (intended to be mesh blocks but can be any other space that's
small enough to be useful. Should also include a column, \code{Person},
with the population within that area.}
}
\value{
a \code{tibble}.
}
\description{
Create a correspondence table between any two geographies.
}
\examples{
make_correspondence_tbl(
  from_geo = get_polygon("sa22016", crs = 7844),
  to_geo = get_polygon("sa22021", crs = 7844),
  mb_geo = mb21_pop
)
}
