% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data_with_correspondence.R
\name{map_data_with_correspondence}
\alias{map_data_with_correspondence}
\title{Map data between editions using correspondence tables from the ABS.}
\usage{
map_data_with_correspondence(
  .data = NULL,
  codes,
  values,
  groups = NULL,
  from_area,
  from_year,
  to_area,
  to_year,
  mb_geo = get_mb21_pop(),
  value_type = c("units", "aggs"),
  round = FALSE,
  seed = NULL
)
}
\arguments{
\item{.data}{a \code{data.frame}(-like) object. Can be passed if codes,
values or groups are passed by reference (like in \code{dplyr::mutate()}).}

\item{codes}{codes representing locations relevant to the \code{from_area}.
SA1 or SA2, for example.}

\item{values}{values associated with codes to be alloocated to newly mapped
codes.}

\item{groups}{values associated with codes/values specifying a grouping
structure. For example, if there are codes/values for several age groups,
then groups will be the grouping variable for the age group associated with
the code and value of the same position. Defaults to \code{NULL} (no
grouping).}

\item{from_area}{The area you want to correspond FROM (ie the areas your data
are currently in). For example: "sa1", "sa2, "sa3", "sa4".}

\item{from_year}{The year you want to correspond FROM. For example: 2011,
2016.}

\item{to_area}{The area you want to correspond TO (ie the areas you want your
data to be in).}

\item{to_year}{The year you want to correspond TO.}

\item{mb_geo}{an \code{{sf}} POINT object where the points are the centroids
of a small area (intended to be mesh blocks but can be any other space that's
small enough to be useful. Should also include a column, \code{Person},
with the population within that area. Defaults to use Mesh Blocks (2021) and
with 2021 census data. See \code{hpa.spatial::mb21_pop}.}

\item{value_type}{Whether the data are unit level or aggregate level data.
Unit level data is randomly allocated to new locations based on proportions
in the correspondence table, aggregate data is dispersed based on the
proportion across the (potentially multiple) mapped codes.}

\item{round}{Whether or not to round the resulting mapped values to be whole
numbers (maybe be useful when mapping count, aggregate values which may
otherwise return decimal values in the mapped areas).}

\item{seed}{A random seed (integer). May be useful for ensuring mappings of
unit level data are reproducible (as these use the mapping probabilities
for randomly allocating observations to the new geographies and may be
different between runs with the same data/inputs).}
}
\value{
A \code{data.frame}.
}
\description{
Map data between editions using correspondence tables from the ABS.
}
\examples{
map_data_with_correspondence(
  codes = c(107011130, 107041149),
  values = c(10, 10),
  from_area = "sa2",
  from_year = 2011,
  to_area = "sa2",
  to_year = 2016,
  value_type = "units"
)
map_data_with_correspondence(
  codes = c(107011130, 107041149),
  values = c(10, 10),
  from_area = "sa2",
  from_year = 2011,
  to_area = "sa2",
  to_year = 2016,
  value_type = "aggs"
)

}
