% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygon.R, R/map_data_with_correspondence.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_absmap}
\alias{get_correspondence_absmaps}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{absmapsdata}{\code{\link[absmapsdata]{get_correspondence_absmaps}}}

  \item{strayr}{\code{\link[strayr]{read_absmap}}}
}}

